Convert ClassicResult implementation to TypeScript
